refactor: only use deck file while deploying
